tpm: Mark region as reserved in case BIOS does not
TPM specification defines one region @0xfed40000-0xfed44fff.  Given a
32bits dom0 with "enough" memory (dom_mem=>4G), it is likely that the
kernel will try to use every last region for RAM, even in upper 32bit
regions.

For some reason, BIOS in HPs or Dells do not report that region in the
e820 as reserved.  It is only reported in one of the SSDT. This patch
tries to work that around by doing so in TBoot.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>