tboot: Unmask NMI potentially masked during SENTER
Add an IRET-to-self entry point ot re-enable NMI and do so right after
re-enabling SMI on BSP and APs. NMIs should be reactivated late to
protect the MLE.

At boot time, Xen 4.11 patches alternatives in NMI context[1] to avoid
NMI or MCE happening while altering their code path. If the self-NMI has
been masked during an SENTER (and apparently it is in some cases), Xen
will panic() with:
    "Timed out waiting for alternatives self-NMI to hit"

Thanks to Ross Philipson at Oracle for his detailed explanations on how
to solve this.

[1] https://github.com/xen-project/xen/commit/5191c1ef51
    5191c1ef51 x86/boot: Make alternative patching NMI-safe

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>