tboot: Add event-log format to exported data.
While TCG format 1.2 and TCG format 2.0 can be differenciated using the
first record of the log, this is more complicated with the legacy Agile TPM
2.0 event-log.

The hash algorithm is passed as part of the heap pointer structure which
does not carry version information. This may not be necessary if the
first element of the event-log is a TPM 1.2 like for TCG format 2.0 and
meaningful data exists in that structure.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>