Default to D/A mapping when TPM1.2 and CBnT platform
TBOOT checks capability bits in SINIT info table to make a decision what PCR
mapping will be used. When both D/A and legacy are supported by SINIT and TPM1.2
is used, TBOOT chose legacy one. This patch modifies that behaviour for CBnT
platforms, so when both mapping are supported, D/A is chosen. For non-CBnT
system nothing is changed.

Signed-off-by: Lukasz Hawrylko <lukasz.hawrylko@intel.com>